+++ docs/changelog.md
@@ -2,6 +2,11 @@
 
 This changelog documents user-relevant changes to the GitHub runner charm.
 
+
+## 2025-02-03
+
+- Add support to integrate with GitHub Runner planner charm with the `github_runner_planner_v0` interface.
+
 ## 2026-02-02
 
 - Deprecated `repo-policy-compliance` service.